+++ src/SpectralAveraging/tst/SpectralIntegration.unit.cpp
@@ -1,6 +1,6 @@
 #include <memory>
 #include <algorithm>
-#include <gtest\gtest.h>
+#include <gtest/gtest.h>
 
 #include "SpectralProperties.hpp"
 #include "IntegratorStrategy.hpp"
@@ -89,7 +89,7 @@ TEST_F( TestSpectralIntegration, TestRectangular )
 
   EXPECT_EQ( calculatedResults.size(), correctResults.size() );
 
-  for( int i = 0; i < calculatedResults.size(); ++i ) {
+  for( size_t i = 0; i < calculatedResults.size(); ++i ) {
     EXPECT_NEAR( correctResults[ i ], calculatedResults[ i ], 1e-6 );
   }
 
@@ -136,7 +136,7 @@ TEST_F( TestSpectralIntegration, TestTrapezoidal )
 
   EXPECT_EQ( calculatedResults.size(), correctResults.size() );
 
-  for( int i = 0; i < calculatedResults.size(); ++i ) {
+  for( size_t i = 0; i < calculatedResults.size(); ++i ) {
     EXPECT_NEAR( correctResults[ i ], calculatedResults[ i ], 1e-6 );
   }
